+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Fix race condition with large indexes (issue #23)
+
 ## [0.6.0] - 2025-06-24
 
 ### Changed